% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonetDBConnection.R
\name{dbConnect}
\alias{dbConnect}
\title{dbConnect}
\usage{
\S4method{dbConnect}{MonetDBDriver}(
  drv,
  dbname = "demo",
  user = "monetdb",
  password = "monetdb",
  host = "localhost",
  port = 50000L,
  timeout = 86400L,
  wait = FALSE,
  language = "sql",
  ...,
  url = ""
)
}
\arguments{
\item{drv}{Should be set to \code{\link[=MonetDB]{MonetDB()}}
to use the \pkg{MonetDB.R} package.}

\item{dbname}{Database name. If \code{NULL}, defaults to the user name.
Note that this argument can only contain the database name, it will not
be parsed as a connection string (internally, \code{expand_dbname} is set to
\code{false} in the call to
\code{\link[=PQconnectdbParams]{PQconnectdbParams()}}(https://www.postgresql.org/docs/9.6/static/    libpq-connect.html)).}

\item{user, password}{User name and password. If \code{NULL}, will be
retrieved from \code{PGUSER} and \code{PGPASSWORD} envvars, or from the
appropriate line in \verb{~/.pgpass}. See
\url{http://www.postgresql.org/docs/9.6/static/libpq-pgpass.html} for
more details.}

\item{host, port}{Host and port. If \code{NULL}, will be retrieved from
\code{PGHOST} and \code{PGPORT} env vars.}

\item{...}{Other name-value pairs that describe additional connection
options as described at
<http://www.postgresql.org/docs/9.6/static/libpq-connect.html#LIBPQ-   PARAMKEYWORDS>}

\item{service}{Name of service to connect as.  If \code{NULL}, will be
ignored.  Otherwise, connection parameters will be loaded from the     pg_service.conf
file and used.  See <http://www.postgresql.org/docs/9.6/static/libpq-  pgservice.html>
for details on this file and syntax.}

\item{check_interrupts}{Should user interrupts be checked during the     query execution (before
first row of data is available)? Setting to \code{TRUE} allows interruption of queries
running too long.}

\item{timezone}{Sets the timezone for the connection. The default is     \code{"UTC"}.
If \code{NULL} then no timezone is set, which defaults to localtime.}

\item{conn}{Connection to disconnect.}
}
\description{
FIXME: the documentation here needs to be updated to MonetDB
\code{DBI::dbConnect()} establishes a connection to a database.
Set \code{drv = MonetDB.R::MonetDB()} to connect to a MonetDB database
using the \pkg{MonetDB.R} package.

Manually disconnecting a connection is not necessary with \pkg{MonetDB.R},
but still recommended;
if you delete the object containing the connection, it will be automatically
disconnected during the next GC with a warning.
}
\examples{
library(DBI)
# Pass more arguments as necessary to dbConnect()
con <- dbConnect(MonetDB.R::MonetDB())
dbDisconnect(con)
}
