% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonetDBConnection.R
\name{dbDataType}
\alias{dbDataType}
\title{dbDataType}
\usage{
\S4method{dbDataType}{MonetDBConnection}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{MonetDB driver or connection.}

\item{obj}{Object to convert}
}
\description{
Determine database type for R vector
}
