% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonetDBConnection.R
\name{dbWriteTable}
\alias{dbWriteTable}
\title{dbWriteTable}
\usage{
\S4method{dbWriteTable}{MonetDBConnection,character}(
  conn,
  name,
  value,
  overwrite = FALSE,
  append = FALSE,
  csvdump = FALSE,
  transaction = TRUE,
  temporary = FALSE,
  ...
)
}
\arguments{
\item{conn}{A MonetDB.R database connection, created using
\code{\link[DBI]{dbConnect}} with the
\code{\link[MonetDB.R]{MonetDB.R}} database driver.}

\item{name}{The name of the database table.}

\item{value}{The dataframe that needs to be stored in the table}

\item{overwrite}{Overwrite the whole table with dataframe. default \code{False}}

\item{append}{Append dataframe to table}

\item{csvdump}{Dump dataframe to a temporary CSV file, and then import that CSV file.
Can be used for performance reasons. Default \code{False}}

\item{transaction}{Wrap operation in transaction. Default: \code{True}}

\item{temporary}{Create a temporary table instead of a 'real' table Default: \code{False}}
}
\value{
TRUE if the writetable command was successful
}
\description{
Write, append or overwrite a data frame to a database table
}
\examples{
dbWriteTable(conn, "mtcars", mtcars[1:5, ])
dbWriteTable(conn, "mtcars", mtcars[5:10, ], overwrite = T)
dbWriteTable(conn, "mtcars", mtcars[11:15, ], append = T)
dbWriteTable(conn, "mtcars", mtcars[11:15, ], append = T, csvdump = T)
dbWriteTable(conn, "iris", iris, temporary = T)
}
