% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{foundDefaultMonetDBdatabase}
\alias{foundDefaultMonetDBdatabase}
\alias{monetdbDefault}
\title{Check if the default MonetDB database is available.}
\usage{
foundDefaultMonetDBdatabase(...)

monetdbDefault(...)
}
\arguments{
\item{...}{Additional arguments passed on to \code{\link[=dbConnect]{dbConnect()}}}
}
\description{
MonetDB examples and tests connect to a default database via
\verb{dbConnect(}\code{\link[=MonetDB]{MonetDB()}}\verb{)}. This function checks if that
database is available, and if not, displays an informative message.

\code{monetdbDefault()} works similarly as \code{foundDefaultMonetDBdatabase()} but
returns a connection on success and throws a testthat skip condition on
failure, making it suitable for use in tests.
}
\examples{
library(DBI)
if (foundDefaultMonetDBdatabase()) {
  conn <- monetdbDefault()
  dbWriteTable(conn, 'mtcars', mtcars)
  dbReadTable(conn, 'mtcars')
  dbRemoveTable(conn, 'mtcars')
  dbListTables(conn)
  dbDisconnect(conn)
}
}
