% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{MonetDB.R-deprecated}
\alias{MonetDB.R-deprecated}
\alias{fetch,MonetDBResult,numeric-method}
\title{Deprecated functions}
\usage{
\S4method{fetch}{MonetDBResult,numeric}(res, n = -1, ...)
}
\arguments{
\item{res}{An object inheriting from \linkS4class{DBIResult}, created by
\code{\link[DBI:dbSendQuery]{dbSendQuery()}}.}

\item{n}{maximum number of records to retrieve per fetch. Use \code{n = -1}
or \code{n = Inf}
to retrieve all pending records.  Some implementations may recognize other
special values.}

\item{...}{Other arguments passed on to methods.}
}
\description{
Deprecated functions kept for backward compatibility purpose.
}
\seealso{
\code{\link[=dbFetch]{dbFetch()}}
}
