% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monetdb_read_csv.R
\name{monetdb.read.csv}
\alias{monetdb.read.csv}
\alias{monet.read.csv}
\title{monetdb.read.csv}
\usage{
monetdb.read.csv(
  conn,
  files,
  tablename,
  header = TRUE,
  best.effort = FALSE,
  delim = ",",
  newline = "\\\\n",
  quote = "\\"",
  create = TRUE,
  col.names = NULL,
  lower.case.names = FALSE,
  sep = delim,
  ...
)

monet.read.csv(
  conn,
  files,
  tablename,
  header = TRUE,
  best.effort = FALSE,
  delim = ",",
  newline = "\\\\n",
  quote = "\\"",
  create = TRUE,
  col.names = NULL,
  lower.case.names = FALSE,
  sep = delim,
  ...
)
}
\arguments{
\item{conn}{A MonetDB.R database connection, created using \link[DBI:dbConnect]{DBI::dbConnect} with
the \link{MonetDB.R} database driver.}

\item{files}{A single string or a vector of strings containing the absolute file
names of the CSV files to be imported.}

\item{tablename}{The dataframe that needs to be stored in the table}

\item{header}{Whether or not the CSV files contain a header line.}

\item{best.effort}{Use best effort flag when reading csv files and continue importing
even if parsing of fields/lines fails.}

\item{delim}{Field separator in the CSV file.}

\item{newline}{Line separator in the CSV file, usually \verb{\\\\n} for UNIX-like systems
and \verb{\\\\r\\\\n} on Windows.}

\item{quote}{Quote character(s) for strings in the CSV file.}

\item{create}{Create table before importing?}

\item{col.names}{Optional column names in case the ones from CSV file should not be
used}

\item{lower.case.names}{Convert all column names to lowercase in the database?}

\item{sep}{Alias for \code{delim}. When specified, overwrite the value of \code{delim}.}

\item{...}{Any other parameters. Ignored.}
}
\value{
NULL if successful; otherwise, error messages
}
\description{
Instruct MonetDB to read a CSV file and use MonetDB's bulk data loading
feature \verb{COPY INTO} to load the CSV data into a database table. By default,
\code{monetdb.read.csv()} also creates the table.

Note that this causes MonetDB to read a file on the machine where the
MonetDB server is running, not on the machine where the R client runs.
}
\examples{
library(DBI)
# Only run the examples on systems with the default MonetDB connection:
if (foundDefaultMonetDBdatabase()) {
  conn <- dbConnect(MonetDB())
  file <- tempfile()
  write.table(iris, file, sep = ",", row.names = FALSE)
  MonetDB.R::monetdb.read.csv(conn, file, "iris")
  DBI::dbGetQuery(conn, paste(
    'SELECT "Species", avg("Sepal.Length"), ',
    '       avg("Sepal.Width"), avg("Petal.Length"), avg("Petal.Width") ',
    'FROM iris GROUP BY "Species"'
  ))
  dbRemoveTable(conn, "iris")
  dbDisconnect(conn)
}
}
\seealso{
\link[utils:read.table]{utils::read.csv}
}
