% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonetDBConnection.R
\docType{class}
\name{MonetDBConnection-class}
\alias{MonetDBConnection-class}
\alias{dbDisconnect,MonetDBConnection-method}
\alias{dbExistsTable,MonetDBConnection,character-method}
\alias{dbGetException,MonetDBConnection-method}
\alias{dbGetInfo,MonetDBConnection-method}
\alias{dbIsValid,MonetDBConnection-method}
\alias{dbListFields,MonetDBConnection,character-method}
\alias{dbListTables,MonetDBConnection-method}
\alias{dbReadTable,MonetDBConnection,character-method}
\alias{dbRemoveTable,MonetDBConnection,character-method}
\title{MonetDBConnection and methods.}
\usage{
\S4method{dbDisconnect}{MonetDBConnection}(conn, ...)

\S4method{dbExistsTable}{MonetDBConnection,character}(conn, name, ...)

\S4method{dbGetException}{MonetDBConnection}(conn, ...)

\S4method{dbGetInfo}{MonetDBConnection}(dbObj, ...)

\S4method{dbIsValid}{MonetDBConnection}(dbObj, ...)

\S4method{dbListFields}{MonetDBConnection,character}(conn, name, ...)

\S4method{dbListTables}{MonetDBConnection}(conn, ..., sys_tables = F, schema_names = F)

\S4method{dbReadTable}{MonetDBConnection,character}(conn, name, ...)

\S4method{dbRemoveTable}{MonetDBConnection,character}(conn, name, ...)
}
\description{
MonetDBConnection and methods.
}
\section{Slots}{

\describe{
\item{\code{connenv}}{<TODO: describe what ever this slot does>}
}}

