% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonetDBDriver.R
\docType{class}
\name{MonetDBDriver-class}
\alias{MonetDBDriver-class}
\alias{dbUnloadDriver,MonetDBDriver-method}
\alias{dbIsValid,MonetDBDriver-method}
\alias{dbGetInfo,MonetDBDriver-method}
\title{MonetDBDriver and methods.}
\usage{
\S4method{dbUnloadDriver}{MonetDBDriver}(drv, ...)

\S4method{dbIsValid}{MonetDBDriver}(dbObj, ...)

\S4method{dbGetInfo}{MonetDBDriver}(dbObj, ...)
}
\description{
MonetDBDriver and methods.
}
