\name{MonetDB.R}
\alias{MonetDB.RDriver}
\alias{MonetDBRDriver}

\title{
  MonetDB.R Driver and methods
}

\usage{
  MonetDB.R ()

  \method{dbIsValid}{MonetDBDriver}(dbObj, ...)
  \method{dbUnloadDriver}{MonetDBDriver}(drv, ...)
  \method{dbGetInfo}{MonetDBDriver}(drv, ...)
  \method{dbConnect}{MonetDBDriver}(dbObj, ...)

}

\keyword{internal}
