% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonetDBResult.R
\docType{class}
\name{MonetDBResult-class}
\alias{MonetDBResult-class}
\alias{fetch,MonetDBResult,numeric-method}
\alias{dbClearResult,MonetDBResult-method}
\alias{dbColumnInfo,MonetDBResult-method}
\alias{dbFetch,MonetDBResult,numeric-method}
\alias{dbGetRowCount,MonetDBResult-method}
\alias{dbGetRowsAffected,MonetDBResult-method}
\alias{dbGetStatement,MonetDBResult-method}
\alias{dbHasCompleted,MonetDBResult-method}
\alias{dbIsValid,MonetDBResult-method}
\title{MonetDBResult and methods.}
\usage{
\S4method{fetch}{MonetDBResult,numeric}(res, n = -1, ...)

\S4method{dbClearResult}{MonetDBResult}(res, ...)

\S4method{dbColumnInfo}{MonetDBResult}(res, ...)

\S4method{dbFetch}{MonetDBResult,numeric}(res, n = -1, ...)

\S4method{dbGetRowCount}{MonetDBResult}(res, ...)

\S4method{dbGetRowsAffected}{MonetDBResult}(res, ...)

\S4method{dbGetStatement}{MonetDBResult}(res, ...)

\S4method{dbHasCompleted}{MonetDBResult}(res, ...)

\S4method{dbIsValid}{MonetDBResult}(dbObj, ...)
}
\description{
MonetDBResult and methods.
}
\section{Slots}{

\describe{
\item{\code{env}}{list of connection environment variables and their values.}
}}

