% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsqlite_compatibility.R
\name{RSQLite-compatibility}
\alias{RSQLite-compatibility}
\alias{isIdCurrent,MonetDBResult-method}
\alias{isIdCurrent,MonetDBConnection-method}
\alias{initExtension,MonetDBConnection-method}
\title{RSQLite Compatibility Functions}
\usage{
\S4method{isIdCurrent}{MonetDBResult}(dbObj, ...)

\S4method{isIdCurrent}{MonetDBConnection}(dbObj, ...)

\S4method{initExtension}{MonetDBConnection}(dbObj, ...)
}
\arguments{
\item{dbObj}{The database object, a \code{MonetDBResult} or \code{MonetDBConnection}.}

\item{...}{Any other parameters. Currently, none is supported.}
}
\description{
Dummy implementations of functions used by RSQLite but not MonetDB so as to
avoid unexpected errors for users who migrate from RSQLite to MonetDB.
}
\seealso{
\code{RSQLite::isIdCurrent()} \code{RSQLite::initExtension()}
}
