% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{dbListTables,MonetDBConnection-method}
\alias{dbListTables,MonetDBConnection-method}
\alias{dbListFields,MonetDBConnection,character-method}
\alias{dbReadTable,MonetDBConnection,character-method}
\alias{dbWriteTable,MonetDBConnection,character-method}
\alias{dbExistsTable,MonetDBConnection,character-method}
\alias{dbRemoveTable,MonetDBConnection,character-method}
\title{Table read/write functions}
\usage{
\S4method{dbListTables}{MonetDBConnection}(conn, ..., sys_tables = F, schema_names = F)

\S4method{dbListFields}{MonetDBConnection,character}(conn, name, ...)

\S4method{dbReadTable}{MonetDBConnection,character}(conn, name, ...)

\S4method{dbWriteTable}{MonetDBConnection,character}(
  conn,
  name,
  value,
  overwrite = FALSE,
  append = FALSE,
  csvdump = FALSE,
  transaction = TRUE,
  temporary = FALSE,
  ...
)

\S4method{dbExistsTable}{MonetDBConnection,character}(conn, name, ...)

\S4method{dbRemoveTable}{MonetDBConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{A MonetDB.R database connection, created using \link[DBI:dbConnect]{DBI::dbConnect} with the
\link{MonetDB.R} database driver.}

\item{...}{Any other parameters. Passed to \code{\link[=monetdb.read.csv]{monetdb.read.csv()}}}

\item{sys_tables}{List the system tables? Default: F}

\item{schema_names}{Include the schema of the tables? Default: F}

\item{name}{The name of the database table.}

\item{value}{The dataframe that needs to be stored in the table}

\item{overwrite}{Overwrite the whole table with dataframe. Default \code{FALSE}.}

\item{append}{Append dataframe to table}

\item{csvdump}{Dump dataframe to a temporary CSV file, and then import that CSV file.
Can be used for performance reasons. Default \code{FALSE}.}

\item{transaction}{Wrap operation in transaction. Default: \code{TRUE}.}

\item{temporary}{Create a temporary table instead of a normal SQL table (i.e.
persistent). Default: \code{FALSE}.}
}
\description{
Functions to read/write data from/into MonetDB SQL tables.

\code{\link[=dbListTables]{dbListTables()}} lists all user created database tables by default. It can
also be instructed to list all system tables and to list all tables with
their schema names.

\code{\link[=dbWriteTable]{dbWriteTable()}} executes several SQL statements in a single transaction to
create/overwrite a database table and fill it with values, or append the
values to an existing database table. It returns a boolean to indicate
success or failure.

\code{\link[=dbRemoveTable]{dbRemoveTable()}} removes the given database table. It returns TRUE to
indicate success removal or FALSE if the table does not exist.
}
\examples{
library(DBI)
if (foundDefaultMonetDBdatabase()) {
  conn <- dbConnect(MonetDB.R())
  dbWriteTable(conn, "mtcars", mtcars[1:5, ])
  dbWriteTable(conn, "mtcars", mtcars[5:10, ], overwrite = TRUE)
  dbWriteTable(conn, "mtcars", mtcars[11:15, ], append = TRUE)
  dbWriteTable(conn, "mtcars", mtcars[11:15, ],
    append = TRUE, csvdump = TRUE
  )
  dbWriteTable(conn, "iris", iris, append = TRUE, temporary = TRUE)
  dbListTables(conn)
  dbListFields(conn, "mtcars")
  dbReadTable(conn, "iris")
  if (dbExistsTable(conn, "iris")) {
    dbRemoveTable(conn, "mtcars")
  }
  dbRemoveTable(conn, "iris")
  dbDisconnect(conn)
}
}
