% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{dbBegin,MonetDBConnection-method}
\alias{dbBegin,MonetDBConnection-method}
\alias{dbCommit,MonetDBConnection-method}
\alias{dbRollback,MonetDBConnection-method}
\title{MonetDB transaction management.}
\usage{
\S4method{dbBegin}{MonetDBConnection}(conn, ...)

\S4method{dbCommit}{MonetDBConnection}(conn, ...)

\S4method{dbRollback}{MonetDBConnection}(conn, ...)
}
\arguments{
\item{conn}{a \linkS4class{MonetDBConnection} object, produced by
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}}

\item{...}{Unused, for extensibility.}
}
\value{
A boolean, indicating success or failure.
}
\description{
Functions to begin, commit or rollback SQL transactions

\code{dbBegin()} starts a transaction. \code{dbCommit()} and \code{dbRollback()}
end the transaction by either committing or rolling back the changes.
}
\examples{
library(DBI)
if (foundDefaultMonetDBdatabase()) {
  conn <- dbConnect(MonetDB.R())

  dbBegin(conn)
  dbWriteTable(conn, "USarrests", datasets::USArrests, temporary = TRUE,
    transaction = FALSE)
  DBI::dbGetQuery(conn, 'SELECT count(*) from "USarrests"')
  dbCommit(conn)

  dbBegin(conn)
  dbExecute(conn, 'DELETE from "USarrests" WHERE "Murder" > 1')
  DBI::dbGetQuery(conn, 'SELECT count(*) from "USarrests"')
  dbRollback(conn)

  # Rolling back changes leads to original count
  DBI::dbGetQuery(conn, 'SELECT count(*) from "USarrests"')

  dbRemoveTable(conn, "USarrests")
  dbDisconnect(conn)
}
}
