% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonetDBConnection.R
\name{dbConnect}
\alias{dbConnect}
\title{dbConnect}
\usage{
\S4method{dbConnect}{MonetDBDriver}(
  drv,
  dbname = "demo",
  user = "monetdb",
  password = "monetdb",
  host = "localhost",
  port = 50000L,
  timeout = 86400L,
  wait = FALSE,
  language = "sql",
  ...,
  url = ""
)
}
\arguments{
\item{drv}{Should be set to \code{\link[=MonetDB]{MonetDB()}}
to use the \pkg{MonetDB.R} package.}

\item{dbname}{The name of the database to connect. Default \code{"demo"}.}

\item{user}{The name of the database user to log in. Default \code{"monetdb"}.}

\item{password}{Password to use. Default \code{"monetdb"}.}

\item{host}{Hostname where the MonetDB server is running.
Default \code{localhost}.}

\item{port}{Port number to which the MonetDB server is listening.
Default: \code{50000}.}

\item{timeout}{Query timeout in seconds. Default: 86400.}

\item{wait}{If the server is not ready yet, wait for it? Default: no.}

\item{language}{The query language to use. Default: \code{"sql"}.}

\item{...}{Other name-value pairs that describe additional connection
options.}

\item{url}{\code{dbname} as a url string. Default: \code{""}.}
}
\value{
conn Connection established to the given MonetDB server.
}
\description{
FIXME: the documentation here needs to be updated to MonetDB
\code{DBI::dbConnect()} establishes a connection to a database.
Set \code{drv = MonetDB.R::MonetDB()} to connect to a MonetDB database
using the \pkg{MonetDB.R} package.

Manually disconnecting a connection is not necessary with \pkg{MonetDB.R},
but still recommended;
if you delete the object containing the connection, it will be automatically
disconnected during the next GC with a warning.
}
\examples{
library(DBI)
# Pass more arguments as necessary to dbConnect()
con <- dbConnect(MonetDB.R::MonetDB())
dbDisconnect(con)
}
