% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonetDBConnection.R
\name{dbDataType,MonetDBConnection-method}
\alias{dbDataType,MonetDBConnection-method}
\title{dbDataType}
\usage{
\S4method{dbDataType}{MonetDBConnection}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{MonetDB driver or connection.}

\item{obj}{R Object to convert}

\item{...}{Any other parameters. Currently, none is supported.}
}
\description{
Determine database type for an R object
}
