% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonetDBConnection.R
\name{dbSendQuery}
\alias{dbSendQuery}
\title{dbSendQuery}
\usage{
\S4method{dbSendQuery}{MonetDBConnection,character}(conn, statement, ..., list = NULL, async = FALSE)
}
\arguments{
\item{conn}{A \linkS4class{MonetDBConnection} created by \code{\link[=dbConnect]{dbConnect()}}.}

\item{statement}{An SQL string to execute}

\item{...}{Another arguments needed for compatibility with generic (currently
ignored).}

\item{list}{A list of extra parameters. Default: \code{NULL}.}

\item{async}{Execute the query in Async mode? Default: \code{FALSE}.}

\item{params}{A list of query parameters to be substituted into
a parameterised query. Query parameters are sent as strings, and the
correct type is imputed by PostgreSQL. If this fails, you can manually
cast the parameter with e.g. \code{"$1::bigint"}.}
}
\description{
FIXME: update this documentation!
Execute a SQL statement on a database connection. This one does all the work
in this class.

To retrieve results a chunk at a time, use \code{dbSendQuery()},
\code{dbFetch()}, then \code{dbClearResult()}. Alternatively, if you want all the
results (and they'll fit in memory) use \code{dbGetQuery()} which sends,
fetches and clears for you.
}
\examples{
library(DBI)
# Only run the examples on systems with the default MonetDB connection:
if (foundDefaultMonetDBdatabase()) {
  conn <- dbConnect(MonetDB.R())
  dbWriteTable(conn, "usarrests", datasets::USArrests, temporary = TRUE)

  # Run query to get results as dataframe
  dbGetQuery(conn, "SELECT * FROM usarrests LIMIT 3")

  # Send query to pull requests in batches
  res <- dbSendQuery(conn, "SELECT * FROM usarrests")
  dbFetch(res, n = 2)
  dbFetch(res, n = 2)
  dbHasCompleted(res)
  dbClearResult(res)
  dbRemoveTable(conn, "usarrests")
  dbDisconnect(conn)
}
}
