% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonetDBConnection.R
\name{dbWriteTable}
\alias{dbWriteTable}
\title{dbWriteTable}
\usage{
\S4method{dbWriteTable}{MonetDBConnection,character}(
  conn,
  name,
  value,
  overwrite = FALSE,
  append = FALSE,
  csvdump = FALSE,
  transaction = TRUE,
  temporary = FALSE,
  ...
)
}
\arguments{
\item{conn}{A MonetDB.R database connection, created using \link[DBI:dbConnect]{DBI::dbConnect} with
the \link{MonetDB.R} database driver.}

\item{name}{The name of the database table.}

\item{value}{The dataframe that needs to be stored in the table}

\item{overwrite}{Overwrite the whole table with dataframe. Default \code{FALSE}.}

\item{append}{Append dataframe to table}

\item{csvdump}{Dump dataframe to a temporary CSV file, and then import that CSV file.
Can be used for performance reasons. Default \code{FALSE}.}

\item{transaction}{Wrap operation in transaction. Default: \code{TRUE}.}

\item{temporary}{Create a temporary table instead of a normal SQL table (i.e.
persistent). Default: \code{FALSE}.}

\item{...}{Any other parameters. Passed to \code{\link[=monetdb.read.csv]{monetdb.read.csv()}}}
}
\value{
TRUE if the writetable command was successful
}
\description{
Write, append or overwrite a data frame to a database table
}
\examples{
library(DBI)
# Only run the examples on systems with the default MonetDB connection:
if (foundDefaultMonetDBdatabase()) {
  conn <- dbConnect(MonetDB.R())
  dbWriteTable(conn, "mtcars", mtcars[1:5, ])
  dbWriteTable(conn, "mtcars", mtcars[5:10, ], overwrite = TRUE)
  dbWriteTable(conn, "mtcars", mtcars[11:15, ], append = TRUE)
  dbWriteTable(conn, "mtcars", mtcars[11:15, ], append = TRUE, csvdump = TRUE)
  dbWriteTable(conn, "iris", iris, append = TRUE, temporary = TRUE)
}
}
