% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonetDBConnection.R
\docType{class}
\name{MonetDBConnection-class}
\alias{MonetDBConnection-class}
\alias{dbGetException,MonetDBConnection-method}
\alias{dbGetInfo,MonetDBConnection-method}
\alias{dbIsValid,MonetDBConnection-method}
\title{MonetDBConnection and methods.}
\usage{
\S4method{dbGetException}{MonetDBConnection}(conn, ...)

\S4method{dbGetInfo}{MonetDBConnection}(dbObj, ...)

\S4method{dbIsValid}{MonetDBConnection}(dbObj, ...)
}
\description{
MonetDBConnection and methods.
}
\section{Slots}{

\describe{
\item{\code{connenv}}{list of connection environment variables and their values.}
}}

\keyword{internal}
