% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonetDBResult.R
\name{RSQLite compatibility}
\alias{RSQLite compatibility}
\alias{isIdCurrent,MonetDBResult-method}
\alias{isIdCurrent,MonetDBConnection-method}
\alias{initExtension,MonetDBConnection-method}
\title{RSQLite Compatibility Functions}
\usage{
\S4method{isIdCurrent}{MonetDBResult}(dbObj, ...)

\S4method{isIdCurrent}{MonetDBConnection}(dbObj, ...)

\S4method{initExtension}{MonetDBConnection}(dbObj, ...)
}
\description{
Functions for RSQLite compatibility.
}
\seealso{
\code{\link[RSQLite:isIdCurrent]{RSQLite::isIdCurrent()}} \code{\link[RSQLite:initExtension]{RSQLite::initExtension()}}
}
